fixed a bug, and added menu disabling for save item
